Refinements to Camera Controller *Added an Area ID field to prevent repeatedly jumping to current starting point. *Changed ints to bytes or sbytes where possible.

+++ Assets/Scripts/Classes/Bullet.cs
@@ -5,7 +5,7 @@ namespace Veganimus.Platformer
     public class Bullet : MonoBehaviour
     {
         [SerializeField] private float _speed = 10.0f;
-        [SerializeField] private int _damageAmount = 1;
+        [SerializeField] private sbyte _damageAmount = 1;
 
         private void Update()
         {
