+++ Assets/Scripts/Scriptable Objects/SceneLoader.cs
@@ -5,9 +5,9 @@ namespace Veganimus.Platformer
     [CreateAssetMenu(menuName = "Scene Loader")]
     public class SceneLoader : ScriptableObject
     {
-        public int sceneIndex;
+        public byte sceneIndex;
 
-        public void LoadGame(int index)
+        public void LoadGame(byte index)
         {
             sceneIndex = index;
             var toUnload = SceneManager.GetActiveScene().buildIndex;
